Lesson #3

+++ Library/src/main/java/com/library/controller/AuthorController.java
@@ -3,43 +3,58 @@ package com.library.controller;
 import com.library.entity.Author;
 import com.library.service.AuthorService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
 import java.util.Optional;
 
 @Controller
-@Component //para que aplicacion functiona he anotado controller con component e etc
-@RequestMapping("/authors")
+@RequestMapping("/api")
 public class AuthorController {
 
     @Autowired
     AuthorService authorService;
 
-    @PostMapping(path = "/POST")
-    public @ResponseBody Author postAuthor (@RequestParam String authorName,
-                                            String dateOfBirth, String nationality){
-        Author newAuthor = authorService.newAuthor(new Author());
+    @PostMapping(path = "/authors")
+    public @ResponseBody Author postAuthor (@RequestBody Author author){
+        Author newAuthor = authorService.newAuthor(author);
         return newAuthor;
     }
-    @DeleteMapping(path = "/DELETE/{id}")
-    public void deleteAuthor(@PathVariable Integer id){
-        authorService.deleteAuthor(id);
-    }
-    @GetMapping(path = "/GET/{id}")
-    public Optional getAuthorWithId(@PathVariable Integer id){
-        Optional authorWithId = authorService.findAuthor(id);
-        if(authorWithId.isPresent()) {
-            return Optional.of(authorWithId);
+    @DeleteMapping(path = "/authors/{id}")
+    public ResponseEntity<HttpStatus> deleteAuthor(@PathVariable Integer id){
+        Optional<Author> idExist = authorService.findAuthor(id);
+        if(idExist.isPresent()) {
+            authorService.deleteAuthor(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }else{
-            return Optional.empty();
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
-    @PutMapping(path = "/PUT/{id}")
-    public Author editAuthor(@PathVariable Integer id, Author author){
-        Author editedAuthor = authorService.editAuthor();//Ni idea que tengo que meter aqui
-        return editedAuthor;
+    @GetMapping(path = "/authors/{id}")
+    @ResponseBody
+    public Author getAuthorWithId(@PathVariable Integer id){
+        Optional<Author> authorWithId = authorService.findAuthor(id);
+        if(authorWithId.isPresent()) {
+            return authorWithId.get();
+        }else{
+            return null;
+        }
+    }
+    @GetMapping(path = "/authors/all")
+    @ResponseBody
+    public List<Author> getAllAuthors(){
+        List<Author> allAuthorsList = authorService.findAllAuthors();
+        return allAuthorsList;
+    }
+    @PutMapping(path = "/authors/{id}")
+    @ResponseBody
+    public Author editAuthor(@PathVariable("id") Integer id, @RequestBody Author author){
+        Optional<Author> editedAuthor = authorService.editAuthor(id, author);
+        return editedAuthor.get();
     }
